First Commit
This project drives an 8 led neo-pixel ring, animating the LEDs to keep your gaze towards your webcam - which you have the lens position in the centre of the ring. So, keep your eye on the lens and thus your audience, rather than staring at the screen and not making eye contaact.

Oh, and by the way this means you can disable the spooky AI in facetime that changes your eyes to make them look at the camera automatically - creepy!

It uses the same psychological insight used by security cameras at Heathrow airport immigration control. The spinning blue LEDs draw your attention subconsciously due to the apparent movement. At the airport this means your eyes are automatically drawn to the security cameras as you walk towards the immigration line!

This is an original Arduino project, not ESP32. It runs on any ATMega328P based Arduino board. I have tested this on the Arduino Uno, but now run it on the Arduino Nano Every board as it is teeny-tiny.

What is the ATMega328p IC?

https://www.microchip.com/en-us/product/atmega328p

What is the Nano Every?

https://store.arduino.cc/products/arduino-nano-every

What the heck is a neopixel?

https://learn.adafruit.com/adafruit-neopixel-uberguide/neopixel-rings

The 8 LED neopixel ring (with 3D printed diffuser)

https://247geek.co.uk/NeopixelRing8Diffuser